fixed rng

+++ src/me/Zachary_Peculier/SIEAHungerGames/Game/Game.java
@@ -67,7 +67,7 @@ public class Game
 
                         for (int amountOfItems = 0; amountOfItems < 9; amountOfItems++)
                         {
-                            double n = Math.random() * 49;
+                            int n = (int) (Math.random() * 49);
                             if (n % 2 == 0)
                             {
                                 inventory.addItem(new ItemStack(randomItens[rand.nextInt(randomItens.length)]));
